+++ projects/_03_independent_cqc/_02_clean/tests/fargate/test_clean_ind_cqc_filled_posts.py
@@ -82,12 +82,10 @@ class MainTests(CleanIndFilledPostsTests):
                 call(
                     ANY,
                     self.CLEANED_IND_CQC_DESTINATION,
-                    append=False,
                 ),
                 call(
                     ANY,
                     self.GROUPED_PROVIDERS_DESTINATION,
-                    append=False,
                 ),
             ]
         )