+++ projects/_03_independent_cqc/_04_model/tests/fargate/test_model_01_features_care_home.py
@@ -53,4 +53,4 @@ class ModelFeaturesCareHomeTests(unittest.TestCase):
         self.assertEqual(cap_integer_at_max_value_mock.call_count, 2)
         self.assertEqual(expand_encode_and_extract_features_mock.call_count, 4)
         select_and_filter_features_data_mock.assert_called_once()
-        sink_to_parquet_mock.assert_called_once_with(ANY, ANY, append=False)
+        sink_to_parquet_mock.assert_called_once_with(ANY, ANY)