+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/tests/fargate/test_02_clean.py
@@ -80,7 +80,6 @@ class MainTests(unittest.TestCase):
         sink_to_parquet_mock.assert_called_once_with(
             lazy_df=ANY,
             output_path=self.CLEANED_DATA_DESTINATION,
-            append=False,
         )
 
     @patch(f"{PATCH_PATH}.utils.sink_to_parquet")