+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/fargate/_02_clean.py
@@ -17,7 +17,6 @@ def main(
     utils.sink_to_parquet(
         lazy_df=lf,
         output_path=cleaned_data_destination,
-        append=False,
     )