+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/utils/test_merge_utils.py
@@ -1,0 +1,6 @@
+import unittest
+from unittest.mock import ANY, Mock, patch
+
+import projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.merge_utils as job
+
+PATCH_PATH = "projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.merge_utils"