+++ CHANGELOG.md
@@ -38,6 +38,10 @@ All notable changes to this project will be documented in this file.
 
 - Changed validate_04_estimates to reflect that historical reallocation is called after manager adjustment.
 
+- Removed `Append` option from `sink_to_parquet` function.
+
+- Added more datasets to SLV merge job.
+
 ### Fixed
 - Replace the all roles and all job groups functions with the appropriate categorical values attribute.